Major changes to program and README, rename files, add new files and content.
