+++ installation_guide.MD
@@ -1,0 +1,170 @@
+# <span style="color:forestgreen">Installation Guide</span>
+
+
+<span style="color:forestgreen">__<-- Back to [README](README.md)__</span>
+
+
+Welcome to the __Installation Guide__! This guide will walk you through installing everything you need to run the __Neat Little Dice Roller__.
+
+
+## Table of Contents
+- [Installation Guide](#installation-guide)
+  - [Table of Contents](#table-of-contents)
+  - [Installation Instructions](#installation-instructions)
+    - [\*\* Developer's Note \*\*](#-developers-note-)
+    - [Installing git, Python 3, pip, and DearPyGui](#installing-git-python-3-pip-and-dearpygui)
+      - [Installing and setting up git](#installing-and-setting-up-git)
+      - [Check if you have Python 3 installed:](#check-if-you-have-python-3-installed)
+      - [Install Python 3:](#install-python-3)
+      - [ pip, or other package installer:](#-pip-or-other-package-installer)
+      - [Install DearPyGui:](#install-dearpygui)
+    - [Installing the Program](#installing-the-program)
+      - [Clone the repository:](#clone-the-repository)
+  - [Running the Program](#running-the-program)
+
+
+## Installation Instructions
+
+The only installations that __Neat Little Dice Roller__  need are git, Python 3, pip (or other package installer), DearPyGui, and of course the actual program files.
+
+
+### <span style="color:firebrick">** Developer's Note **</span>
+
+<span style="color:firebrick">_I developed this program using MacOS. The language in this README may reflect that such as "Terminal" vs "Command Prompt", or "cmd+C" vs "ctrl+C", etc, despite my best efforts to include both... __Please just substitute the appropriate command in it's place!___</span> :smiley:
+
+- <span style="color:firebrick">The __Terminal__ (MacOS) or __Command Prompt__ (Windows) are where you type in the command line commands.</span>
+
+- <span style="color:firebrick">Keyboard shortcuts often use __cmd+C__ (cmd = command)(MacOS) or __ctrl+C__ (ctrl = control)(Windows).</span>
+  - <span style="color:firebrick">_Note: Replace "C" with whatever shortcut is needed._</span>
+
+
+### Installing git, Python 3, pip, and DearPyGui
+
+#### <span style="color:limegreen">Installing and setting up git</span>
+
+You will have a different method of installation, depending on your operating system.
+
+__<u>Check if git is installed:</u>__
+
+- Open the __Terminal__ or __Command Prompt__
+  - Type `git --version`
+    - If git is installed correctly, the command will display the installed version.
+
+
+__<u>Install:</u>__
+
+_<span style="color:firebrick">If you already have git installed you can skip this section!</style>_
+
+- <span style="color:mediumaquamarine">Windows:</span>
+  - Download the installer from the Git website and run it.
+    - https://git-scm.com/downloads/win
+
+- <span style="color:mediumaquamarine">macOS:</span>
+    - Open the __Terminal__
+
+    - Type `brew install git`
+      - If for some reason homebrew is not installed, install it here:
+        - https://brew.sh/
+
+
+__<u>Configure:</u>__
+- Set your username:
+  - Type `git config --global user.name "Your Name"`
+
+- Set your email:
+  - Type `git config -- global user.email "your.email@example.com"`
+
+
+#### <span style="color:limegreen">Check if you have Python 3 installed:</span>
+
+- Open the __Terminal__ or __Command Prompt__
+
+- Type `python3 --version`
+  - If Python 3 is installed correctly, the command will display the installed version
+
+
+#### <span style="color:limegreen">Install Python 3:</span>
+
+_<span style="color:firebrick">If you already have Python 3 installed you can skip this section!</style>_
+
+- Go to the official Python 3 downloads page:
+  - https://www.python.org/downloads/
+    - It should have the appropriate version for your computer on this page. 
+
+  - Click the yellow button labeled __Download Python 3.13.5__ 
+    - This is the current version as of creation of this README. Your version may be higher, depending on when you've found this!
+
+  - This should download to your downloads folder (or elsewhere depending on your computer settings.)
+
+- Run the installer:
+  - Find your downloaded installer, and run it.
+
+    - <span style="color:red">__Important:__ If you are on Windows, make sure to check the box labeled __Add Python to PATH__</span>
+
+- When the installation is finished, verify your installation:
+  - We'll do this ths same way we checked for a Python 3 installation in the first place:
+    - Open the __Terminal__ or __Command Prompt__
+
+    - Type `python3 --version`
+      - If Python 3 is installed correctly, the command will display the installed version
+
+
+#### <span style="color:limegreen"> pip, or other package installer:</span>
+
+<span style="color:firebrick">_The next step after this one involves using a package installer. Most Python 3 installations come with pip already installed, and my instructions use pip. Obviously feel free to pick the one you are most comfortable with._</span> 
+
+- Open the __Terminal__ or __Command Prompt__
+
+- Type `pip --version`
+  - It will give you the currently installed version of pip.
+
+If pip is not already installed, you can find instructions elsewhere online. One place is pip's documentation:
+- https://pip.pypa.io/en/stable/installation/
+
+
+#### <span style="color:limegreen">Install DearPyGui:</span>
+
+- Open the __Terminal__ or __Command Prompt__
+  - Type `pip install dearpygui`
+
+- Check that the installation worked:
+  - Type `pip show dearpygui`
+    - This should show you information about the installed dearpygui package
+
+
+### Installing the Program
+
+#### <span style="color:limegreen">Clone the repository:</span>
+
+- Navigate to the main page of this repository:
+  - https://github.com/XLove3/NLDR
+
+- Above the list of files, click the button __<> Code__
+
+- Copy the url for the repository (using HTTPS):
+  - Make sure you are on the __Local__ tab, and __HTTPS__ is the underlined method of copying
+
+  - Click on the copy button (two overlapping squares) to the right of the URL
+
+- Open __Terminal__ or __Command Prompt__
+
+- Navigate to the file location where you would like this repository to live using `cd`
+
+- Type `git clone`, and then paste the URL you copied earlier
+
+    `git clone https://github.com/XLove3/NLDR.git`
+
+- Press enter to create your local copy of the program files
+
+
+*****
+*****
+
+## Running the Program
+
+To run the program: 
+- Open the __Terminal__ or __Command Prompt__
+
+- Navigate to the folder holding this program using `cd`
+
+- Type `python3 main.py`
